requirements.txt
selenium
webdriver-manager
requests
beautifulsoup4
imap-tools
python-dotenv
rich